+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
+## <next>
+* Count of selectable items added to multiselect label
+* Upgraded @opuscapita/react-floating-select version
+
 ## 8.0.1
 * Fix header cell sort click, check deeper into child tree